refactor to match what cdk init generates
