+++ .gitignore
@@ -1,2 +1,3 @@
 /target
 /deployment/node_modules/
+/cdk/node_modules/
