+++ cdk/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "cdk",
+  "version": "0.1.0",
+  "bin": {
+    "cdk": "bin/cdk.js"
+  },
+  "scripts": {
+    "build": "tsc",
+    "watch": "tsc -w",
+    "test": "jest",
+    "cdk": "cdk"
+  },
+  "devDependencies": {
+    "@aws-cdk/core": "1.143.0",
+    "@aws-cdk/aws-apigatewayv2": "1.143.0",
+    "@aws-cdk/aws-apigatewayv2-integrations": "1.143.0",
+    "@aws-cdk/aws-dynamodb": "1.143.0",
+    "@aws-cdk/aws-iam": "1.143.0",
+    "@aws-cdk/aws-lambda": "1.143.0",
+    "@types/jest": "^26.0.10",
+    "@types/node": "17.0.15",
+    "jest": "^26.4.2",
+    "ts-jest": "^26.2.0",
+    "aws-cdk": "2.10.0",
+    "ts-node": "10.4.0",
+    "typescript": "4.5.5"
+  },
+  "dependencies": {
+    "aws-cdk-lib": "2.15.0",
+    "constructs": "^10.0.0",
+    "source-map-support": "^0.5.16"
+  }
+}
