--- deployment/cdk.json
@@ -1,3 +1,0 @@
-{
-    "app": "npx ts-node Deploy.ts"
-}
