--- deployment/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "cdk",
-  "version": "1.0.0",
-  "description": "",
-  "scripts": {
-    "build": "ts"
-  },
-  "devDependencies": {
-    "@aws-cdk/core": "1.143.0",
-    "@aws-cdk/aws-apigatewayv2": "1.143.0",
-    "@aws-cdk/aws-apigatewayv2-integrations": "1.143.0",
-    "@aws-cdk/aws-dynamodb": "1.143.0",
-    "@aws-cdk/aws-iam": "1.143.0",
-    "@aws-cdk/aws-lambda": "1.143.0",
-    "@types/node": "17.0.15",
-    "aws-cdk": "2.10.0",
-    "parcel-bundler": "1.12.5",
-    "typescript": "4.5.5",
-    "ts-node": "10.4.0"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
-}
