--- deployment/tsconfig.json
@@ -1,31 +1,0 @@
-{
-    "compilerOptions": {
-        "outDir": "dist",
-        "target": "es6",
-        "useDefineForClassFields": true,
-        "lib": [
-            "DOM",
-            "DOM.Iterable",
-            "ESNext",
-            "es2015",
-            "es2017",
-        ],
-        "allowJs": false,
-        "skipLibCheck": false,
-        "esModuleInterop": true,
-        "allowSyntheticDefaultImports": true,
-        "strict": false,
-        "forceConsistentCasingInFileNames": true,
-        "module": "commonjs",
-        "moduleResolution": "Node",
-        "resolveJsonModule": true,
-        "isolatedModules": true,
-        "noEmit": true,
-    },
-    "include": [
-        "src/**/*"
-    ],
-    "exclude": [
-        "node_modules/**/*"
-    ]
-}
